+++ sdk/eventhub/event-hubs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 5.0.3 (Unreleased)
 
+- Fixes an issue where known properties were not being set on a received event's `systemProperties`.
 
 ## 5.0.2 (2020-03-09)